+++ src/main/scala/parser/nodes.scala
@@ -1,33 +1,17 @@
 package parser
 
+import lexer.TokenStream
 import org.joda.time.DateTime
+import parser.ast.functions.ClassMethodDecl
 
-sealed trait LangNode
-
-sealed trait ImportClause
-case class ImportAlias(name:String, alias:Option[String])
-case class FromCljRequire(from: String, names: List[ImportAlias]) extends ImportClause
-case class FromJvmRequire(from: String, names: List[ImportAlias]) extends ImportClause
-case class CljImport(name:List[ImportAlias]) extends ImportClause
-case class JvmImport(name:List[ImportAlias]) extends ImportClause
-
-case class Module(name: String, imports: Option[List[ImportClause]],decls: List[LangNode]) extends LangNode
-
-case class TraitMethodDecl(name: String, args: List[String])
-case class TraitDecl(inner: Boolean, name: String, decls: List[TraitMethodDecl]) extends LangNode
-
-case class ClassMethodDecl(definition: DefDecl)
-case class TraitDef(traitName: String, methods: List[ClassMethodDecl]) extends LangNode
-case class ClassDecl(inner: Boolean, name: String, args: Option[List[String]], traits: List[TraitDef]) extends LangNode
+trait LangNode
 
 
-case class RecordDecl(name: String, args: List[String]) extends LangNode
-
-case class ExtendsDecl(cls: String, traitClass: String, decls: Option[List[ClassMethodDecl]]) extends LangNode
 
 
-case class ADT(name: String, args: List[String])
-case class AdtDecl(name: String, defs: List[ADT]) extends LangNode
+
+
+
 trait AssignableExpression
 
 
@@ -36,7 +20,6 @@ sealed trait Expression extends LangNode
 case class ReifyExpression(traitName: String, methods: List[ClassMethodDecl]) extends Expression
 
 
-case class TopLevelExpression(expression: Expression) extends LangNode
 
 class Name(name: String) extends Expression
 trait LambdaArg
@@ -77,11 +60,6 @@ case class WhereBlock(whereDefs: List[WhereDef]) extends LangNode
 case class DefArg(expression: Expression)
 object DefOtherwiseArg extends DefArg(null)
 
-
-sealed trait DispatcherTrait
-object ClassDispatcher extends DispatcherTrait
-case class ExpressionDispatcher(expr:Expression) extends DispatcherTrait
-case class DispatchDecl(id:String, dispatcher: DispatcherTrait) extends LangNode
 
 class DefDecl(id: String) extends LangNode
 case class MultiMethod(inner: Boolean, id: String, matches: List[DefArg], args: List[DefArg], body: Expression, whereBlock: Option[WhereBlock])
@@ -141,8 +119,6 @@ case class FStringLiteral(value: String) extends LiteralExpression
 trait CallExpression extends Expression
 case class FunctionCallExpression(func: Expression, args:List[Expression]) extends CallExpression
 case class FunctionCallWithDollarExpression(func: Expression, args: List[Expression]) extends CallExpression
-case class ForwardPipeFuncCallExpression(args: List[Expression]) extends CallExpression
-case class ForwardPipeFirstArgFuncCallExpression(args: List[Expression]) extends CallExpression
 case class BackwardPipeFuncCallExpression(args: List[Expression]) extends CallExpression
 case class BackwardPipeFirstArgFuncCallExpression(args: List[Expression]) extends CallExpression
 case class NewCallExpression(cls: String, args: List[Expression]) extends CallExpression